requirements.txt
langchain
langchain-chroma
langchain-community
langchain-core
langchain-openai
pypdf
unstructured
unstructured-inference
unstructured-pdf-image
striprtf
python-dotenv
transformers
torch
sentence-transformers
accelerate
bitsandbytes
huggingface-hub
scipy
numpy
